--- automation/src/test/scala/org/broadinstitute/dsde/firecloud/fixture/UISpec.scala
@@ -1,62 +1,0 @@
-package org.broadinstitute.dsde.firecloud.fixture
-
-import org.broadinstitute.dsde.workbench.auth.AuthToken
-import org.broadinstitute.dsde.workbench.config.{Credentials, UserPool}
-import org.broadinstitute.dsde.workbench.fixture._
-import org.broadinstitute.dsde.workbench.service.test.{RandomUtil, WebBrowserSpec}
-import org.scalatest.{BeforeAndAfterAll, Matchers, Outcome, fixture}
-
-import scala.util.Try
-
-
-abstract class UISpec extends fixture.FreeSpec with BeforeAndAfterAll with BillingFixtures with Matchers
-  with WebBrowserSpec with UserFixtures with WorkspaceFixtures with GroupFixtures with RandomUtil {
-
-  val owner: Credentials = UserPool.chooseProjectOwner
-  implicit val ownerAuthToken: AuthToken = owner.makeAuthToken()
-  var claimedBillingProject: ClaimedProject = _
-
-  /**
-    * See
-    *  https://www.artima.com/docs-scalatest-2.0.M5/org/scalatest/FreeSpec.html
-    *   Section: "Overriding withFixture(OneArgTest)"
-    *
-    * @param billingProject
-    */
-  case class projectOwnerFixture(billingProject: String)
-
-  type FixtureParam = projectOwnerFixture
-
-  override def withFixture(test: OneArgTest): Outcome = {
-    withFixture(test.toNoArgTest(projectOwnerFixture(claimedBillingProject.projectName)))
-  }
-
-  override def beforeAll(): Unit = {
-    Try {
-      claimedBillingProject = claimGPAllocProject(owner)
-      logger.info(s"beforeAll(): billing project claimed: ${claimedBillingProject.projectName}")
-      super.beforeAll()
-    }.recover {
-      case ex: Exception =>
-        logger.error("Error occurred in billing project claim", ex)
-        throw ex // ends test
-    }
-  }
-
-  override def afterAll(): Unit = {
-    try {
-      super.afterAll()
-    } finally {
-      val projectName = claimedBillingProject.projectName
-      Try {
-        claimedBillingProject.cleanup(owner, List())
-        logger.info(s"afterAll(): billing project cleaned: $projectName")
-      }.recover{
-        case ex: Exception =>
-          logger.error(s"Error occurred in billing project clean - $projectName", ex)
-          throw ex
-      }
-
-    }
-  }
-}
